Move NDErrorData and NDModuleError to common/exceptions.py
