+++ plugins/module_utils/nd_v2.py
@@ -50,116 +50,13 @@ import logging
 from typing import Any, Optional
 
 from ansible.module_utils.basic import env_fallback
-from ansible_collections.cisco.nd.plugins.module_utils.common.pydantic_compat import BaseModel, ConfigDict
+from ansible_collections.cisco.nd.plugins.module_utils.common.exceptions import NDModuleError
 from ansible_collections.cisco.nd.plugins.module_utils.enums import HttpVerbEnum
 from ansible_collections.cisco.nd.plugins.module_utils.rest.protocols.response_handler import ResponseHandlerProtocol
 from ansible_collections.cisco.nd.plugins.module_utils.rest.protocols.sender import SenderProtocol
 from ansible_collections.cisco.nd.plugins.module_utils.rest.response_handler_nd import ResponseHandler
 from ansible_collections.cisco.nd.plugins.module_utils.rest.rest_send import RestSend
 from ansible_collections.cisco.nd.plugins.module_utils.rest.sender_nd import Sender
-
-
-class NDErrorData(BaseModel):
-    """
-    # Summary
-
-    Pydantic model for structured error data from NDModule requests.
-
-    This model provides type-safe error information that can be serialized
-    to a dict for use with Ansible's fail_json.
-
-    ## Attributes
-
-    - msg: Human-readable error message (required)
-    - status: HTTP status code as integer (optional)
-    - request_payload: Request payload that was sent (optional)
-    - response_payload: Response payload from controller (optional)
-    - raw: Raw response content for non-JSON responses (optional)
-    """
-
-    model_config = ConfigDict(extra="forbid")
-
-    msg: str
-    status: Optional[int] = None
-    request_payload: Optional[dict[str, Any]] = None
-    response_payload: Optional[dict[str, Any]] = None
-    raw: Optional[Any] = None
-
-
-class NDModuleError(Exception):
-    """
-    # Summary
-
-    Exception raised by NDModule when a request fails.
-
-    This exception wraps an NDErrorData Pydantic model, providing structured
-    error information that can be used by callers to build appropriate error
-    responses (e.g., Ansible fail_json).
-
-    ## Usage Example
-
-    ```python
-    try:
-        data = nd.request("/api/v1/endpoint", HttpVerbEnum.POST, payload)
-    except NDModuleError as e:
-        print(f"Error: {e.msg}")
-        print(f"Status: {e.status}")
-        if e.response_payload:
-            print(f"Response: {e.response_payload}")
-        # Use to_dict() for fail_json
-        module.fail_json(**e.to_dict())
-    ```
-    """
-
-    def __init__(
-        self,
-        msg: str,
-        status: Optional[int] = None,
-        request_payload: Optional[dict[str, Any]] = None,
-        response_payload: Optional[dict[str, Any]] = None,
-        raw: Optional[Any] = None,
-    ) -> None:
-        self.error_data = NDErrorData(
-            msg=msg,
-            status=status,
-            request_payload=request_payload,
-            response_payload=response_payload,
-            raw=raw,
-        )
-        super().__init__(msg)
-
-    @property
-    def msg(self) -> str:
-        """Human-readable error message."""
-        return self.error_data.msg
-
-    @property
-    def status(self) -> Optional[int]:
-        """HTTP status code."""
-        return self.error_data.status
-
-    @property
-    def request_payload(self) -> Optional[dict[str, Any]]:
-        """Request payload that was sent."""
-        return self.error_data.request_payload
-
-    @property
-    def response_payload(self) -> Optional[dict[str, Any]]:
-        """Response payload from controller."""
-        return self.error_data.response_payload
-
-    @property
-    def raw(self) -> Optional[Any]:
-        """Raw response content for non-JSON responses."""
-        return self.error_data.raw
-
-    def to_dict(self) -> dict[str, Any]:
-        """
-        Convert exception attributes to a dict for use with fail_json.
-
-        Returns a dict containing only non-None attributes.
-        """
-        return self.error_data.model_dump(exclude_none=True)
 
 
 def nd_argument_spec() -> dict[str, Any]:
